# backend/src/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Analysis API",
    "version": "1.0.0",
    "description": "API for analyzing queries using AI"
  },
  "servers": [
    {
      "url": "https://pulse-ai.fly.dev",
      "description": "Production server"
    }
  ],
  "paths": {
    "/api/analyze": {
      "post": {
        "summary": "Analyze a query",
        "operationId": "analyzeQuery",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["query"],
                "properties": {
                  "query": {
                    "type": "string",
                    "description": "The query to analyze"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful analysis",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "answer": {
                      "type": "string",
                      "description": "The analysis result"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "description": "Error message",
                      "example": "Failed to process query"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/addTask": {
      "post": {
        "summary": "Create a Notion task",
        "operationId": "addTask",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "data": {
                    "type": "object",
                    "description": "Any JSON data structure to create a task from"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Task created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "link": {
                      "type": "string",
                      "description": "The Notion page URL for the created task"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "description": "Error message",
                      "example": "Failed to create task"
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  }
}